NXP-26746: Add kafka SSL SASL to the stack
This is pretty hard to deploy Kafka/Zookeeper in SSL in a test
environment and especially in docker-compose.
Because it is a test env the certificate needs to be self signed and the
hostname verification must be tweaked, this is possible on the Kafka
side using an empty kafka.ssl.endpoint.identification.algorithm, but
from the Nuxeo side we need to pass a dns check to resolve the kafka
hostname, a way to do this is using a static IP for kafka along with an
extr_hosts docker compose option that translates into a /etc/hosts entry.
The jaas file is shared between Zookeeper (Server section),
Kafka (KafkaServer section) and Nuxeo (Client section).
There is no keystore on the Nuxeo side because of the SASL PLAIN auth.